interview phase: rename configuration to settings
